refactored such that server.js calls the ChatRoom and TCP server and that all tests work
