+++ server.js
@@ -1,31 +1,7 @@
-const net = require('net');
-const ChatRoom = require('./lib/Chatroom');
-const chatRoom = new ChatRoom();
+const server = require('./lib/tcpServer');
 
-const server = net.createServer(client => {
-    client.setEncoding('utf-8');
-    var messageParsed = '';
-    var newNick = '';
-
-    chatRoom.add(client);
-
-    //chatRoom.whoshere();
-
-    client.on('data', message => {
-        chatRoom.send(client, message);
-        messageParsed = message.split(' ');
-        if (messageParsed[0] === '/nick' && 1 < messageParsed.length) {
-            newNick = messageParsed[1].replace('\r\n','');
-            chatRoom.changeName(client, newNick);
-        };
-        console.log(`${client.name}: ${message}`);
-    });
-
-    client.on('close', () => {
-        chatRoom.disconnectMessage(client);
-        chatRoom.remove(client);
-        console.log(`${client.name} has disconnected`);
-    });
+const port = 65000;
+server.listen(port, err => {
+    if (err) console.log('Error! ', err);
+    else console.log('Listening to port ', port);
 });
-
-module.exports = server;
